Export to and import from directory
Storing the contents of all `Zettel`s in the same file is a bad idea.
As this file will grow very quickly. So `Zettel`s will be stored as
individual files.

In preparation for that the tool is now able to export to a directory
and import from a directory.