Fix 'ZettelDoesntExistsError' when creating new Zettels
The problem was raised because the `App` is not in sync with the content
on the file system.

If a new Zettel is created it is added to `App`. Next we try open the
Zettel with an editor. Because the `App` hasn't been exported yet the
Zettel also hasn't been exported to the file system. So trying to open
the Zettel with an editor fails.

The quick solution is to export the `App` to disk after the Zettel has
been added. But before the Zettel is opened. This is not very nice as it
requires 2 calls to `App.export()` when creating a single Zettel.